Rename files to prevent collisions and fallbacks
