+++ packages/openapi/test/amazonIon.test.ts
@@ -3,12 +3,12 @@ import { createHTTPHandler } from '@trpc/server/adapters/standalone';
 import { makeAsyncResource } from '@trpc/server/unstable-core-do-not-import/stream/utils/disposable';
 import { describe, expect, it } from 'vitest';
 import { configureTRPCHeyApiClient } from '../src/heyapi';
+import { Sdk } from './routers/amazonIonRouter-heyapi';
+import { client } from './routers/amazonIonRouter-heyapi/client.gen';
 import {
   AmazonIonRouter,
   amazonIonTransformer,
-} from './routers/amazonIonRouter';
-import { Sdk } from './routers/amazonIonRouter-heyapi';
-import { client } from './routers/amazonIonRouter-heyapi/client.gen';
+} from './routers/amazonIonRouter.router';
 
 const richInput = {
   name: 'Alice',
