+++ packages/openapi/test/routers/cyclicTypesRouter.openapi.json
