+++ packages/openapi/test/routers/edgeCaseRouter.openapi.json
