+++ packages/openapi/test/routers/edgeCaseRouter.router.ts
