+++ packages/openapi/test/routers/errorFormatterRouter.openapi.json
